requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
python-multipart==0.0.9
stripe==10.12.0
redis==5.0.8
groq==0.11.0
chromadb==0.5.3
sentence-transformers==3.0.1
sse-starlette==2.1.3
httpx==0.27.2

